devops_soft-main/forStep4/requirements.txt
Flask==3.0.3
mysql-connector-python==8.2.0
